infinite loop - stable

+++ hw2_test.cxx
@@ -1,16 +1,16 @@
 #include "hw2_test.h"
 
-int set_status(int status) {
+long set_status(int status) {
     long r = syscall(334, status);
     return r;
 }
 
-int get_status() {
+long get_status() {
     long r = syscall(335);
     return r;
 }
 
-int register_process() {
+long register_process() {
 	long r = syscall(336);
     return r;
 }
