+++ hw2_test.h
@@ -6,9 +6,9 @@
 #include <sys/syscall.h>
 
 // System call wrappers
-int set_status(int status);
-int get_status();
-int register_process();
+long set_status(int status);
+long get_status();
+long register_process();
 long get_all_cs();
 
 #endif // test_hw2_H_
